dcf1: Fixed bug in debug facility for synchonization process.
We had forgotten to add a #ifdef, #define, #endif triad in our driver
code that then used to define the symbol that is used to enable debug
output for the respective change.

In order to make sure this will not happen again, I remove the ifdef
triad block completly. From now on the symbols generated by Kconfig
(prefixed CONFIG_) will be used to control the debug macros.